+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Sort terms in `BakeChapterGlossary` in language specific way (major)
 
 ## [7.0.0] - 2021-06-21